Move and rename tagApi.ts
This is to be in accordance with the recent changes to file structure and naming conventions surrounding API files.
